Move SearchProgressListener to correct package
