+++ MultiSubDownloader/src/main/java/org/lodder/subtools/multisubdownloader/listeners/SearchProgressListener.java
@@ -1,6 +1,5 @@
-package org.lodder.subtools.multisubdownloader.gui.dialog.progress.search;
+package org.lodder.subtools.multisubdownloader.listeners;
 
-import org.lodder.subtools.multisubdownloader.listeners.StatusListener;
 import org.lodder.subtools.multisubdownloader.subtitleproviders.SubtitleProvider;
 import org.lodder.subtools.sublibrary.model.Release;
